Bug 2196: Remove test from MetaType CT.
The removed test was in TestControl.testBundleMetaTypeProvider() where it was required that a MetaTypeService provider ensure the underlying
class of the OCD retrieved from the service was the same class as the one from the MetaTypeProvider. There is no obvious reason for this
requirement. Implementations should be allowed to use their own OCD classes, if desired, as long as they remain faithful to the data provided
by the original OCD class.