[Bug 2487] Change AD default value in CT to match one of the options.
Previously, the CT used an empty string as the default value for an attribute definition with specified
options. None of the options was an empty string. The specification is unclear what an implementation
should do in this circumstance. For example, are options inclusive or exclusive? May/Must/Should an
implementation validate the specified default value? If so, should just the default value be rejected? The
attribute definition? The object class definition? The entire XML?

Note that this change does not affect the fidelity of the CT. The intent was simply to verify that an attribute
definition returned the specified (and valid) default value, not to test the corner case where the default
value does not match one of the options.